# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  handlers.AvailabilityResponse:
    properties:
      data:
        properties:
          is_available:
            example: false
            type: boolean
        type: object
      success:
        example: true
        type: boolean
    type: object
...
  handlers.ErrorResponse:
    properties:
      error:
        example: invalid request
        type: string
      success:
        example: false
        type: boolean
    type: object
...
  handlers.MessageResponse:
    properties:
      data:
        additionalProperties:
          type: string
        type: object
      success:
        example: true
        type: boolean
    type: object
  handlers.PropertyImageResponse:
    properties:
      data:
        $ref: '#/definitions/models.PropertyImage'
      success:
        example: true
        type: boolean
    type: object
  handlers.PropertyListResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/models.Property'
        type: array
      page:
        example: 1
        type: integer
      page_size:
        example: 10
        type: integer
      pages:
        example: 10
        type: integer
      success:
        example: true
        type: boolean
      total:
        example: 100
        type: integer
    type: object
  handlers.PropertyResponse:
    properties:
      data:
        $ref: '#/definitions/models.Property'
      success:
        example: true
        type: boolean
    type: object
...
  models.Property:
    properties:
      average_rating:
        description: Computed (not stored)
        type: number
      bathrooms:
        type: integer
      city:
        type: string
      country:
        description: Location — from broad to precise
        type: string
      created_at:
        type: string
      currency:
        type: string
      description:
        type: string
      exact_address:
        type: string
      furnished:
        description: Amenities
        type: boolean
      has_courtyard:
        type: boolean
      has_electricity:
        type: boolean
      has_water:
        type: boolean
      has_wifi:
        type: boolean
      id:
        type: integer
      images:
        items:
          $ref: '#/definitions/models.PropertyImage'
        type: array
      is_available:
        type: boolean
      is_favorited:
        type: boolean
      latitude:
        type: number
      longitude:
        type: number
      neighborhood:
        description: sub-area within the city (optional)
        type: string
      owner:
        allOf:
        - $ref: '#/definitions/models.User'
        description: Relations
      owner_id:
        type: integer
      price:
        description: Pricing — ISO 4217 currency codes (KES, USD, EUR, GBP, NGN, GHS,
          TZS, UGX …)
        type: number
      property_type:
        description: Bedsitter, Studio, Apartment, Maisonette, Bungalow, Townhouse,
          Villa, Commercial
        type: string
      review_count:
        type: integer
      reviews:
        items:
          $ref: '#/definitions/models.Review'
        type: array
      rooms:
        description: Size & layout
        type: integer
      shower_type:
        description: en_suite, shared
        type: string
      state:
        description: state / county / region / province (optional)
        type: string
      surface:
        description: m²
        type: number
      title:
        description: Listing info
        type: string
      transaction_type:
        description: rent, sale
        type: string
      updated_at:
        type: string
    type: object
  models.PropertyImage:
    properties:
      created_at:
        type: string
      filename:
        type: string
      id:
        type: integer
      is_main:
        type: boolean
      property_id:
        type: integer
    type: object
...
  models.Review:
    properties:
      comment:
        type: string
      created_at:
        type: string
      id:
        type: integer
      is_verified:
        type: boolean
      property_id:
        type: integer
      rating:
        description: 1-5
        type: integer
      user:
        $ref: '#/definitions/models.User'
      user_id:
        type: integer
    type: object
...
  models.User:
    properties:
      active:
        type: boolean
      created_at:
        type: string
      date_of_birth:
        type: string
      email:
        type: string
      first_name:
        type: string
      gender:
        type: string
      id:
        type: integer
      last_name:
        type: string
      locale:
        type: string
      member_since:
        type: string
      phone_number:
        type: string
      profile_picture:
        type: string
      timezone:
        type: string
      updated_at:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@gidana.com
    name: Gidana Support
  description: Real-estate platform API — property listings, rentals, messaging, digital
    wallets & payments.
  title: Gidana API
  version: 1.0.0
paths:
...
  /images/{id}:
    delete:
      parameters:
      - description: Image ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.MessageResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete a property image
      tags:
      - images
  /images/{id}/main:
    patch:
      parameters:
      - description: Image ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.MessageResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Set an image as the main property photo
      tags:
      - images
  /properties:
    get:
      parameters:
      - description: Free-text search
        in: query
        name: q
        type: string
      - description: Country
        in: query
        name: country
        type: string
      - description: City
        in: query
        name: city
        type: string
      - description: Property type (Apartment, Studio, Bedsitter…)
        in: query
        name: property_type
        type: string
      - description: rent or sale
        in: query
        name: transaction_type
        type: string
      - description: Minimum price
        in: query
        name: min_price
        type: number
      - description: Maximum price
        in: query
        name: max_price
        type: number
      - description: Page number (default 1)
        in: query
        name: page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.PropertyListResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: List properties with optional filters
      tags:
      - properties
    post:
      consumes:
      - multipart/form-data
      parameters:
      - description: Title
        in: formData
        name: title
        required: true
        type: string
      - description: Description
        in: formData
        name: description
        type: string
      - description: Bedsitter | Studio | Apartment | Maisonette | Bungalow | Townhouse
          | Villa | Commercial
        in: formData
        name: property_type
        required: true
        type: string
      - description: rent | sale
        in: formData
        name: transaction_type
        required: true
        type: string
      - description: Country
        in: formData
        name: country
        required: true
        type: string
      - description: City
        in: formData
        name: city
        required: true
        type: string
      - description: State / region
        in: formData
        name: state
        type: string
      - description: Neighborhood
        in: formData
        name: neighborhood
        type: string
      - description: Exact address
        in: formData
        name: exact_address
        type: string
      - description: Latitude
        in: formData
        name: latitude
        type: number
      - description: Longitude
        in: formData
        name: longitude
        type: number
      - description: Number of rooms
        in: formData
        name: rooms
        required: true
        type: integer
      - description: Number of bathrooms
        in: formData
        name: bathrooms
        required: true
        type: integer
      - description: en_suite | shared
        in: formData
        name: shower_type
        type: string
      - description: Surface area in m²
        in: formData
        name: surface
        type: number
      - description: Is furnished
        in: formData
        name: furnished
        type: boolean
      - description: Has Wi-Fi
        in: formData
        name: has_wifi
        type: boolean
      - description: Has water
        in: formData
        name: has_water
        type: boolean
      - description: Has electricity
        in: formData
        name: has_electricity
        type: boolean
      - description: Has courtyard
        in: formData
        name: has_courtyard
        type: boolean
      - description: Price
        in: formData
        name: price
        required: true
        type: number
      - description: ISO 4217 currency code (e.g. XAF, USD)
        in: formData
        name: currency
        required: true
        type: string
      - description: Property images (multiple)
        in: formData
        name: images
        type: file
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handlers.PropertyResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a new property listing
      tags:
      - properties
  /properties/{id}:
    delete:
      parameters:
      - description: Property ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.MessageResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete a property listing
      tags:
      - properties
    get:
      parameters:
      - description: Property ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.PropertyResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Get a single property by ID
      tags:
      - properties
    put:
      consumes:
      - multipart/form-data
      parameters:
      - description: Property ID
        in: path
        name: id
        required: true
        type: integer
      - description: Title
        in: formData
        name: title
        required: true
        type: string
      - description: Description
        in: formData
        name: description
        type: string
      - description: Property type
        in: formData
        name: property_type
        required: true
        type: string
      - description: rent | sale
        in: formData
        name: transaction_type
        required: true
        type: string
      - description: Country
        in: formData
        name: country
        required: true
        type: string
      - description: City
        in: formData
        name: city
        required: true
        type: string
      - description: State / region
        in: formData
        name: state
        type: string
      - description: Neighborhood
        in: formData
        name: neighborhood
        type: string
      - description: Exact address
        in: formData
        name: exact_address
        type: string
      - description: Latitude
        in: formData
        name: latitude
        type: number
      - description: Longitude
        in: formData
        name: longitude
        type: number
      - description: Number of rooms
        in: formData
        name: rooms
        required: true
        type: integer
      - description: Number of bathrooms
        in: formData
        name: bathrooms
        required: true
        type: integer
      - description: en_suite | shared
        in: formData
        name: shower_type
        type: string
      - description: Surface area in m²
        in: formData
        name: surface
        type: number
      - description: Is furnished
        in: formData
        name: furnished
        type: boolean
      - description: Has Wi-Fi
        in: formData
        name: has_wifi
        type: boolean
      - description: Has water
        in: formData
        name: has_water
        type: boolean
      - description: Has electricity
        in: formData
        name: has_electricity
        type: boolean
      - description: Has courtyard
        in: formData
        name: has_courtyard
        type: boolean
      - description: Price
        in: formData
        name: price
        required: true
        type: number
      - description: ISO 4217 currency code
        in: formData
        name: currency
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.PropertyResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update a property listing
      tags:
      - properties
  /properties/{id}/availability:
    patch:
      parameters:
      - description: Property ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.AvailabilityResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Toggle a property's availability
      tags:
      - properties
  /properties/{id}/images:
    post:
      consumes:
      - multipart/form-data
      parameters:
      - description: Property ID
        in: path
        name: id
        required: true
        type: integer
      - description: Image file
        in: formData
        name: image
        required: true
        type: file
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handlers.PropertyImageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Add an image to a property
      tags:
      - properties
...
  /properties/featured:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.PropertyListResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Get featured properties
      tags:
      - properties
  /properties/my/listings:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.PropertyListResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List the current user's property listings
      tags:
      - properties
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and your JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"